kafka_scripts/requirements.txt
kafka-python==2.0.2
faker==20.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
pandas==2.1.4
